Added custom Query and Repository class
Added a new method with a custom Query, very simple for now. It was only
added to the interface, still need to implement it to be able to call
it.

Also added a new Interface with a custom Repository, it doesn't do
anything, it was just to test the functionality for now. Might implement
it later.